module-3/experiments/requirements.txt
pandas==2.2.2
numpy==1.26.4
lancedb==0.12.0
wandb==0.17.9
librosa==0.10.2.post1
torch==2.4.1
transformers==4.44.2
python-dotenv==1.0.1